Fix/gc sync yaw pitch (#208)
* MP2 GC: Add camera_state memory address

* MP1 GC: sync pitch when camera is not FPS-controlled

* MP2 GC: sync yaw/pitch when camera is not FPS-controlled

When the game temporarily owns the camera (morph ball, cutscenes), PrimeHack continued to use stale yaw/pitch state. This caused the view to snap or point in the wrong direction when FPS control resumed.

* MP2 GC: sync yaw/pitch fix for cutscenes

The sync of yaw/pitch must happen for one more frame in order to work with cutscenes. Was not able to test all.